Move "create_operation" from CSV row to its own class
